notebooks/3-sm-job/requirements.txt
tensorflow-hub>=0.8.0
tensorflow-text==1.15.1
tensorflow-gpu==1.15.3
